refactor(tracing): extract context watcher helpers
