+++ tests/internal/events/test_context_watcher.py
@@ -15,8 +15,8 @@ pytestmark = pytest.mark.skipif(
 
 @pytest.fixture(autouse=True)
 def _register_context_watcher():
-    from ddtrace.internal.native._native import is_context_watcher_registered
-    from ddtrace.internal.native._native import register_context_watcher
+    from ddtrace.internal._context_watcher import is_context_watcher_registered
+    from ddtrace.internal._context_watcher import register_context_watcher
 
     assert register_context_watcher()
     assert is_context_watcher_registered()
@@ -108,8 +108,8 @@ def test_context_watcher_slot_exhaustion_disables_watcher():
         raise AssertionError("context watcher slots were not exhausted")
 
     try:
-        from ddtrace.internal.native._native import is_context_watcher_registered
-        from ddtrace.internal.native._native import register_context_watcher
+        from ddtrace.internal._context_watcher import is_context_watcher_registered
+        from ddtrace.internal._context_watcher import register_context_watcher
 
         assert register_context_watcher() is False
         assert is_context_watcher_registered() is False
@@ -127,8 +127,8 @@ def test_context_watcher_registration_is_idempotent():
     from contextvars import Context
 
     from ddtrace.internal import core
-    from ddtrace.internal.native._native import is_context_watcher_registered
-    from ddtrace.internal.native._native import register_context_watcher
+    from ddtrace.internal._context_watcher import is_context_watcher_registered
+    from ddtrace.internal._context_watcher import register_context_watcher
 
     assert is_context_watcher_registered() is False
